Started created the template for the code
Mainly cleared out any starting information that is not required.  Also,
note that low power programming is implemented.... at the end of the
main loop, HLT flag will trigger "Halt Mode".  This will be released by
the WDT function.  Testing using a WDT timing of 8sec.  We will need to
optimize for the program use and this WDT timing.